add dependency to fast-channel and add control to control ssh key
