+++ web/package.json
@@ -56,8 +56,7 @@
     "redux-form": "^7.4.2",
     "redux-thunk": "^2.3.0",
     "sha.js": "^2.4.11",
-    "spark-md5": "^3.0.1",
-    "video-headline-fast-channels": "git+ssh://git@github.com:qualabs/video-headline-fast-channels.git#hotfix/add-package-json"
+    "spark-md5": "^3.0.1"
   },
   "devDependencies": {
     "eslint-plugin-flowtype": "^2.39.1",
@@ -66,7 +65,7 @@
     "react-addons-perf": "^15.4.2"
   },
   "scripts": {
-    "preinstall": "node check-ssh-key.js",
+    "preinstall": "node fast-channel-dependency.js",
     "build-css": "node-sass-chokidar --include-path ./src --include-path ./node_modules src/ -o src/",
     "watch-css": "npm run build-css && node-sass-chokidar --include-path ./src --include-path ./node_modules src/ -o src/ --watch --recursive",
     "start-js": "react-scripts start",
