+++ web/src/components/OperatorSite/navigation.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {Switch, Route, Link} from 'react-router-dom'
 import history from '../../history'
-
+import { FASTChannelDashboard, CreateNewChannel, CreateAdConfiguration, Scheduler } from 'video-headline-fast-channels/dist'
 import DashboardOperator from './DashboardOperator'
 import VideoList from './VideoList'
 import VideoDetail from './VideoDetail'
@@ -21,23 +21,22 @@ import {clearVideo} from '../../actions/video'
 import LiveDetail from './LiveDetail'
 import BillDetail from './BillDetail'
 import Monitor from './Monitor'
-//let FASTChannelDashboard, CreateNewChannel, CreateAdConfiguration, Scheduler;
 
-// if (isSshKeyAvailable()) {
-//   ({ FASTChannelDashboard, CreateNewChannel, CreateAdConfiguration, Scheduler } =import('video-headline-fast-channels'));
-// }
+const isSshKeyAvailable = require('../../check-ssh-key'); 
 
-import FASTChannelDashboard from 'video-headline-fast-channels'
+
+
 export class OperatorRoutes extends Component {
   constructor (props) {
     super(props)
-    //this.config = this.props.user.organization.config
+    this.config = this.props.user.organization.config
   }
   render () {
     return (
       <Switch>
-        {/* /* {this.config && FASTChannelDashboard && (
-        <Route exact path='/fast-channels/' render={(props) => <FASTChannelDashboard history={history}
+        {isSshKeyAvailable && (
+          <Route exact path='/fast-channels/' render={(props) => 
+          <FASTChannelDashboard history={history}
           routes={{newChannel: '/fast-channels/channel/new',
             newAd: '/fast-channels/ad/new',
             scheduler: '/fast-channels/scheduler'}}
@@ -45,20 +44,18 @@ export class OperatorRoutes extends Component {
         />}
         />
         )}
-        {this.config && FASTCCreateNewChannelhannelDashboard && ( <Route exact path='/fast-channels/channel/new/' component={CreateNewChannel}/>)}
-        {this.config && CreateAdConfiguration && (<Route exact path='/fast-channels/ad/new/' component={CreateAdConfiguration}/>)}
-        {this.config && Scheduler &&  (<Route exact path='/fast-channels/scheduler/:id/'  render={(props) =>
-          isSshKeyAvailable() ? <Scheduler history={history} {...props} /> : <Redirect to='/' />
-        }/> )} */ }
-
-         <Route exact path='/fast-channels/' render={(props) => <FASTChannelDashboard history={history}
-          routes={{newChannel: '/fast-channels/channel/new',
-            newAd: '/fast-channels/ad/new',
-            scheduler: '/fast-channels/scheduler'}}
-          {...props}
-          />} />
-
-  
+        {isSshKeyAvailable && ( 
+          <Route exact path='/fast-channels/channel/new/' component={CreateNewChannel}/>
+        )}
+        {isSshKeyAvailable && (
+          <Route exact path='/fast-channels/ad/new/' component={CreateAdConfiguration}/>
+        )}
+        {isSshKeyAvailable  &&( 
+          <Route exact path='/fast-channels/scheduler/:id/' render={() => 
+            <Scheduler history={history}
+          />}
+          /> 
+        )}
 
         <Route exact path='/video-on-demand/new/' component={VideoNew} />
         <Route exact path='/video-on-demand/:id/*/' component={VideoDetail} />
